+++ lib/core/database/luna_database.dart
@@ -29,6 +29,7 @@ class LunaDatabase extends LunaModuleDatabase {
                 case LunaDatabaseValue.QUICK_ACTIONS_SONARR:
                 case LunaDatabaseValue.QUICK_ACTIONS_NZBGET:
                 case LunaDatabaseValue.QUICK_ACTIONS_SABNZBD:
+                case LunaDatabaseValue.QUICK_ACTIONS_OMBI:
                 case LunaDatabaseValue.QUICK_ACTIONS_TAUTULLI:
                 case LunaDatabaseValue.QUICK_ACTIONS_SEARCH:
                 case LunaDatabaseValue.USE_24_HOUR_TIME:
@@ -61,6 +62,7 @@ class LunaDatabase extends LunaModuleDatabase {
                 case LunaDatabaseValue.QUICK_ACTIONS_NZBGET:
                 case LunaDatabaseValue.QUICK_ACTIONS_SABNZBD:
                 case LunaDatabaseValue.QUICK_ACTIONS_TAUTULLI:
+                case LunaDatabaseValue.QUICK_ACTIONS_OMBI:
                 case LunaDatabaseValue.QUICK_ACTIONS_SEARCH:
                 case LunaDatabaseValue.USE_24_HOUR_TIME:
                 case LunaDatabaseValue.LATEST_CHANGELOG_BUILD_SHOWN: value.put(config[key]); break;
@@ -85,6 +87,7 @@ class LunaDatabase extends LunaModuleDatabase {
             case 'LUNASEA_QUICK_ACTIONS_SONARR': return LunaDatabaseValue.QUICK_ACTIONS_SONARR;
             case 'LUNASEA_QUICK_ACTIONS_NZBGET': return LunaDatabaseValue.QUICK_ACTIONS_NZBGET;
             case 'LUNASEA_QUICK_ACTIONS_SABNZBD': return LunaDatabaseValue.QUICK_ACTIONS_SABNZBD;
+            case 'LUNASEA_QUICK_ACTIONS_OMBI': return LunaDatabaseValue.QUICK_ACTIONS_OMBI;
             case 'LUNASEA_QUICK_ACTIONS_TAUTULLI': return LunaDatabaseValue.QUICK_ACTIONS_TAUTULLI;
             case 'LUNASEA_QUICK_ACTIONS_SEARCH': return LunaDatabaseValue.QUICK_ACTIONS_SEARCH;
             case 'LUNASEA_USE_24_HOUR_TIME': return LunaDatabaseValue.USE_24_HOUR_TIME;
@@ -110,6 +113,7 @@ enum LunaDatabaseValue {
     QUICK_ACTIONS_SONARR,
     QUICK_ACTIONS_NZBGET,
     QUICK_ACTIONS_SABNZBD,
+    QUICK_ACTIONS_OMBI,
     QUICK_ACTIONS_TAUTULLI,
     QUICK_ACTIONS_SEARCH,
     USE_24_HOUR_TIME,
@@ -135,6 +139,7 @@ extension LunaDatabaseValueExtension on LunaDatabaseValue {
             case LunaDatabaseValue.QUICK_ACTIONS_NZBGET: return 'LUNASEA_QUICK_ACTIONS_NZBGET';
             case LunaDatabaseValue.QUICK_ACTIONS_SABNZBD: return 'LUNASEA_QUICK_ACTIONS_SABNZBD';
             case LunaDatabaseValue.QUICK_ACTIONS_TAUTULLI: return 'LUNASEA_QUICK_ACTIONS_TAUTULLI';
+            case LunaDatabaseValue.QUICK_ACTIONS_OMBI: return 'LUNASEA_QUICK_ACTIONS_OMBI';
             case LunaDatabaseValue.QUICK_ACTIONS_SEARCH: return 'LUNASEA_QUICK_ACTIONS_SEARCH';
             case LunaDatabaseValue.USE_24_HOUR_TIME: return 'LUNASEA_USE_24_HOUR_TIME';
             case LunaDatabaseValue.LATEST_CHANGELOG_BUILD_SHOWN: return 'LUNASEA_LATEST_CHANGELOG_BUILD_SHOWN';
@@ -161,6 +166,7 @@ extension LunaDatabaseValueExtension on LunaDatabaseValue {
             case LunaDatabaseValue.QUICK_ACTIONS_NZBGET: return box.get(this.key, defaultValue: false);
             case LunaDatabaseValue.QUICK_ACTIONS_SABNZBD: return box.get(this.key, defaultValue: false);
             case LunaDatabaseValue.QUICK_ACTIONS_TAUTULLI: return box.get(this.key, defaultValue: false);
+            case LunaDatabaseValue.QUICK_ACTIONS_OMBI: return box.get(this.key, defaultValue: false);
             case LunaDatabaseValue.QUICK_ACTIONS_SEARCH: return box.get(this.key, defaultValue: false);
             case LunaDatabaseValue.USE_24_HOUR_TIME: return box.get(this.key, defaultValue: false);
             case LunaDatabaseValue.LATEST_CHANGELOG_BUILD_SHOWN: return box.get(this.key, defaultValue: null);
@@ -186,6 +192,7 @@ extension LunaDatabaseValueExtension on LunaDatabaseValue {
             case LunaDatabaseValue.QUICK_ACTIONS_SONARR: if(value.runtimeType == bool) box.put(this.key, value); return;
             case LunaDatabaseValue.QUICK_ACTIONS_NZBGET: if(value.runtimeType == bool) box.put(this.key, value); return;
             case LunaDatabaseValue.QUICK_ACTIONS_SABNZBD: if(value.runtimeType == bool) box.put(this.key, value); return;
+            case LunaDatabaseValue.QUICK_ACTIONS_OMBI: if(value.runtimeType == bool) box.put(this.key, value); return;
             case LunaDatabaseValue.QUICK_ACTIONS_TAUTULLI: if(value.runtimeType == bool) box.put(this.key, value); return;
             case LunaDatabaseValue.QUICK_ACTIONS_SEARCH: if(value.runtimeType == bool) box.put(this.key, value); return;
             case LunaDatabaseValue.USE_24_HOUR_TIME: if(value.runtimeType == bool) box.put(this.key, value); return;