+++ lib/modules/radarr/core/router.dart
@@ -5,6 +5,8 @@ import 'package:lunasea/modules/radarr.dart';
 class RadarrRouter extends LunaModuleRouter {
     @override
     void defineAllRoutes(FluroRouter router) {
-        RadarrHomeRouter.defineRoutes(router);
+        RadarrHomeRouter().defineRoute(router);
+        RadarrMoviesAddRouter().defineRoute(router);
+        RadarrMoviesDetailsRouter().defineRoute(router);
     }
 }