+++ lib/modules/radarr/modules.dart
@@ -3,5 +3,6 @@ export 'modules/missing.dart';
 export 'modules/movies.dart';
 export 'modules/movies_add.dart';
 export 'modules/movies_add_details.dart';
+export 'modules/movies_details.dart';
 export 'modules/radarr.dart';
 export 'modules/upcoming.dart';