+++ lib/modules/radarr/modules/movies_details.dart
@@ -1,0 +1,2 @@
+export 'movies_details/route.dart';
+export 'movies_details/widgets.dart';