+++ lib/modules/radarr/modules/radarr/widgets.dart
@@ -1,2 +1,4 @@
+export 'widgets/appbar_add_movie_action.dart';
+export 'widgets/appbar_global_settings_action.dart';
 export 'widgets/appbar.dart';
 export 'widgets/navigation_bar.dart';