+++ lib/modules/sabnzbd/core/constants.dart
@@ -1,26 +1,23 @@
 import 'package:flutter/material.dart';
 import 'package:lunasea/core.dart';
+import 'package:lunasea/modules/sabnzbd.dart';
 
 class SABnzbdConstants {
     SABnzbdConstants._();
 
     static const String MODULE_KEY = 'sabnzbd';
 
-    static const LunaModuleMetadata MODULE_METADATA = LunaModuleMetadata(
+    //ignore: non_constant_identifier_names
+    static LunaModuleMetadata MODULE_METADATA = LunaModuleMetadata(
         name: 'SABnzbd',
         description: 'Manage Usenet Downloads',
         settingsDescription: 'Configure SABnzbd',
         helpMessage: 'SABnzbd is a multi-platform binary newsgroup downloader. The program works in the background and simplifies the downloading verifying and extracting of files from Usenet.',
         icon: CustomIcons.sabnzbd,
-        route: '/sabnzbd',
+        route: SABnzbd.ROUTE_NAME,
         color: Color(0xFFFECC2B),
         website: 'https://sabnzbd.org',
         github: 'https://github.com/sabnzbd/sabnzbd',
-    );
-
-    //ignore: non_constant_identifier_names
-    static final ShortcutItem MODULE_QUICK_ACTION = ShortcutItem(
-        type: MODULE_KEY,
-        localizedTitle: MODULE_METADATA.name,
+        shortcutItem: ShortcutItem(type: MODULE_KEY, localizedTitle: 'SABnzbd'),
     );
 }